validateStatusMsgRequest now take Temporary Visible contacts into account
git-svn-id: http://miranda.googlecode.com/svn/trunk@1678 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/protocols/IcqOscarJ/utilities.h
@@ -105,7 +105,7 @@ BOOL writeDbInfoSettingByteWithTable(HANDLE hContact, const char *szSetting, str
 
 int GetGMTOffset(void);
 
-BOOL validateStatusMessageRequest(HANDLE hContact, BYTE byMessageType);
+BOOL validateStatusMessageRequest(HANDLE hContact, WORD byMessageType);
 
 #define icqOnline ((gnCurrentStatus != ID_STATUS_OFFLINE) && (gnCurrentStatus != ID_STATUS_CONNECTING))
